Summary: Fixed error from cat when trying nginx.d/* on nodes without vhosts defined
Description:
When running first time from scratch, if class was included without any setup at all
and with sane defaults, in debian squeeze generates an error when trying to collect
non existant vhosts, because cat complains that can not expand * because there are no
files.

Changed:
Added unless to check that at least there is a file in nginx.d by testing it can expand.