sort $vhost_cfg_append hash in vhost_footer.erb template
This is so the paremter ordering in the configuration is stable. It also
force the :allow key (if it exists) to be first in the sort ordering.
It would probably be better to change `$vhost_cfg_append` into an array
of one key hashes but would require a public API change.