Revert "Added ngnix::resources::upstream::member"
Upstream members can no longer be exported and collected.

The change in #331 was fundamentally broken. I have therefore reverted
it as it shouldn't of been merged.

Essentially you can't use ensure with this change - meaning you can no
longer REMOVE an nginx config from the system - which is part of the
tests and also sane module practice.

The idea was nice - but the implementation broke things. This reverts
back to a good state, without modifying any tests where tests pass again
with the recent commits.

This reverts commit af6cda249e94f00d724eae60e4325f9cb5ccdb71.