Disable proxy_http_version directive by default
The default value in nginx::config was the same as nginx's default,
so it's not necessary to be set in any version of nginx unless one
needs a specific value. This also allows us to use undef when we don't
want the directive to be set instead of false.

Also added a test for the default case.