apt::key: puppetlabs-apt check now the full GPG fingerprints.
Since commit 531ef9 of puppetlabs-apt, the gpg key (apt::key) is
checked, and should use the long format.

This value is validated with a regex enforcing it to only contain valid
hexadecimal characters, be precisely 8 or 16 hexadecimal characters long
and optionally prefixed with 0x for key IDs, or 40 hexadecimal
characters long for key fingerprints.

https://github.com/puppetlabs/puppetlabs-apt/commit/f588f2651a68c5863aac7dac910d96bbc7531ef9