Manage the nginx log directory.
Ensure the nginx::config::log_dir exists so that the access.log and error.log can be created. Also any nginx::vhost also put their access/error logs in this directory as well. This is done so that on platforms where /var/log/nginx is not created by package installation it exists so that nginx can startup and create the log files.

When vhost purge is enabled the log directory will not be touched and the log files will remain.