Fix dependency on apt-transport-https
Without this change, the nginx https repository url is added to sources.list.d
and apt-get update is executed immediately after, failing because
apt-transport-https is not yet installed. Subsequently all package
installs fail because the apt cache has been invalidated.

With this change we ensure the apt-transport-https package is present
before running apt-get update. Since all sources are now https, the
package is always necessary, which is why I moved it out of the 'case'
statement.