Replace `add_listen_directive` with `nginx_version`
Nginx 1.16 has recently been released.  Instead of updating the README
to recommend users set the `add_listen_directive` parameter, this
parameter has been replaced by an `nginx_version` parameter that
defaults to the fact (if available).  If the fact isn't available, it
defaults to a very conservative `1.6.0` (the version that ships in Debian
8).  Future enhancements could make the non-fact default be OS specific.