Simplify service class
It's valid to pass in undef values. That makes the OpenBSD check
redundant. The same goes for the restart command. That removes
pretty much all logic so there is no more benefit to having class
parameters. That's why it now uses all the variables directly from the
nginx class. The has{status,restart} parameters default to true since
Puppet 2.7 so can be removed.