fix documentation about ipv6only always being added
This was refering to issue #30 which was closed by changing the
behavior for *not* adding the default in the template. Instead, the
default includes the setting. We do not mention the default here
because it's visible in the generated `REFERENCE.md` file and is not
mentioned in other parameter documentation anyways.